# bradobrei_party/backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API для системы управления сетью барбершопов Bradobrei Party.",
        "title": "Bradobrei Party API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/inventories/salon/{salonId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает фактические остатки материалов на складе выбранного салона.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Остатки материалов по салону",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID салона",
                        "name": "salonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Inventory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inventories/salon/{salonId}/material/{materialId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Устанавливает фактическое количество материала на складе конкретного салона по ID материала.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventories"
                ],
                "summary": "Обновить остаток материала на складе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID салона",
                        "name": "salonId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID материала",
                        "name": "materialId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое количество материала",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SetInventoryQuantityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Inventory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "bad_request"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "dto.SetInventoryQuantityRequest": {
            "type": "object",
            "required": [
                "quantity"
            ],
            "properties": {
                "quantity": {
                    "type": "number",
                    "minimum": 0,
                    "example": 25
                }
            }
        },
...
        "models.EmployeeProfile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expected_salary": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "salons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Salon"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "specialization": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "work_schedule": {
                    "description": "{\"mon\":\"9-18\",\"tue\":\"9-18\",...}",
                    "type": "string"
                }
            }
        },
        "models.Inventory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "last_updated": {
                    "type": "string"
                },
                "material": {
                    "$ref": "#/definitions/models.Material"
                },
                "material_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "salon": {
                    "$ref": "#/definitions/models.Salon"
                },
                "salon_id": {
                    "type": "integer"
                }
            }
        },
        "models.Material": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "description": "мл, шт, гр",
                    "type": "string"
                }
            }
        },
...
        "models.Salon": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "base_hourly_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EmployeeProfile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "description": "PostGIS, опционально",
                    "type": "string"
                },
                "max_staff": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "OPEN/CLOSED",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "working_hours": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EmployeeProfile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServiceMaterial"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ServiceMaterial": {
            "type": "object",
            "properties": {
                "material": {
                    "$ref": "#/definitions/models.Material"
                },
                "material_id": {
                    "type": "integer"
                },
                "quantity_per_use": {
                    "type": "number"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "service_id": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employee_profile": {
                    "$ref": "#/definitions/models.EmployeeProfile"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "CLIENT",
                "BASIC_MASTER",
                "ADVANCED_MASTER",
                "HR_SPECIALIST",
                "ACCOUNTANT",
                "NETWORK_MANAGER",
                "ADMINISTRATOR"
            ],
            "x-enum-varnames": [
                "RoleClient",
                "RoleBasicMaster",
                "RoleAdvancedMaster",
                "RoleHR",
                "RoleAccountant",
                "RoleNetworkManager",
                "RoleAdmin"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}